refactor(F241): extract agentProvider manifest parser
Why: Keep plugin-manifest.ts under the repo line-limit convention before Slice 2b/2c add more agentProvider surface, while preserving the reviewed Slice 2a parser behavior.

Validation: pnpm --filter @cat-cafe/api run build; focused agentProvider/providerTransport tests 85/85; pnpm --filter @cat-cafe/api run lint; pnpm check; git diff --check.

[砚砚/gpt-5.5🐾]
